// frontend/src/api/client/models/ParcoursDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ParcoursDto = {
    id: string;
    slug: string;
    name: string;
    description: string;
    defis_per_period: number;
    quizz_per_period: number;
    period_type: string;
};